Create README.md
We are a team of chemical and computer science engineers dedicated to creating innovative web tools. Our platform offers practical chemical engineering solutions with online calculators to simplify complex calculations. As hobbyists, we aim to make intricate problems more accessible and easier to solve through the power of coding and technology.